requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.11.12
